API difficulty fix to remove Itoa2 for cross compatibility
This quick API update help to share  reward, difficulty and network_hashrate.   It include the removing of the Itoa2 to keep the real number of the difficulty for prevening bug on cointomine when the pool api is listed.
It also add the network hashrate to yiimp based on the current difficulty and Time To Find.
Note : The modification of ApiController.php not need reboot.